Update Shogun's DTC benchmark
Renamed the older file to dtc.py
Also made the newer version support more configurable options than the
older version.
	deleted:    methods/shogun/decision_tree.py
	new file:   methods/shogun/dtc.py
